+++ src/clj/salava/admin/helper.clj
@@ -4,7 +4,9 @@
             [clojure.java.jdbc :as jdbc]
             [salava.core.helper :refer [dump]]
             [slingshot.slingshot :refer :all]
-            [salava.core.util :refer [get-db]]))
+            [salava.core.util :refer [get-db]]
+            [clojure.data.csv :as csv]
+            [clojure.java.io :as io]))
 
 (defqueries "sql/admin/queries.sql")
 
@@ -13,3 +15,8 @@
   [ctx user-id]
   (let [admin (select-user-admin {:id user-id} (into {:result-set-fn first :row-fn :role} (get-db ctx)))]
     (= admin "admin")))
+
+(defn make-csv [ctx data]
+ (fn [out]
+  (with-open [writer (io/writer out)]
+   (csv/write-csv writer data :separator \:))))