+++ src/clj/salava/badge/endorsement.clj
@@ -1,13 +1,15 @@
 (ns salava.badge.endorsement
   (:require [yesql.core :refer [defqueries]]
-            [salava.core.util :refer [get-db md->html get-full-path plugin-fun get-plugins event publish]]
+            [salava.core.util :refer [get-db md->html get-full-path plugin-fun get-plugins event publish digest bytes->base64]]
             [slingshot.slingshot :refer :all]
             [clojure.tools.logging :as log]
             [salava.badge.main :refer [send-badge-info-to-obf badge-exists?]]
-            [salava.core.time :as time]))
+            [salava.core.time :as time]
+            [salava.badge.ext-endorsement :as ext]))
 
 (defqueries "sql/badge/main.sql")
 (defqueries "sql/badge/endorsement.sql")
+(defqueries "sql/badge/ext_endorsement.sql")
 
 (defn generate-external-id []
   (str "urn:uuid:" (java.util.UUID/randomUUID)))
@@ -90,7 +92,7 @@
 
 (defn user-badge-endorsements
   ([ctx user-badge-id]
-   (select-user-badge-endorsements {:user_badge_id user-badge-id} (get-db ctx)))
+   (concat (select-user-badge-endorsements {:user_badge_id user-badge-id} (get-db ctx)) (map #(assoc % :type "ext" )(ext/select-user-badge-ext-endorsements {:user_badge_id user-badge-id} (get-db ctx)))))
   ([ctx user-badge-id html?]
    (reduce (fn [r e]
              (conj r (-> e (update :content md->html)))) [] (user-badge-endorsements ctx user-badge-id))))
@@ -99,7 +101,14 @@
   (select-user-badge-endorsements-p {:id user-badge-id} (get-db ctx)))
 
 (defn received-pending-endorsements [ctx user-id]
-  (map (fn [e] (-> e (update :content md->html))) (select-pending-endorsements {:user_id user-id} (get-db ctx))))
+ (->> (list*
+       (map (fn [e] (-> e (update :content md->html))) (select-pending-endorsements {:user_id user-id} (get-db ctx)))
+       (some->> (ext/endorsements-received ctx user-id true)
+                (filter #(= (:status %) "pending"))
+                (map #(update % :content md->html))
+                (map #(assoc % :type "ext"))))
+      flatten
+      (sort-by :mtime >)))
 
 (defn endorsements-received
  ([ctx user-id]
@@ -131,11 +140,16 @@
         given (endorsements-given ctx user-id)
         requests (->> (endorsement-requests ctx user-id) (filterv #(= (:status %) "pending")) (mapv #(assoc % :type "request")))
         sent-requests (some->> (select-sent-endorsement-requests {:id user-id} (get-db ctx)) (mapv #(assoc % :type "sent_request")))
-        all (->> (list* given received requests sent-requests) flatten (sort-by :mtime >))]
+        ext-received (some->> (if md? (ext/endorsements-received ctx user-id true)(ext/endorsements-received ctx user-id))
+                              (mapv #(assoc % :type "ext")))
+        ext-sent-requests (some->> (select-sent-ext-endorsement-requests {:id user-id} (get-db ctx)) (mapv #(assoc % :type "ext_request")))
+        all (->> (list* given received requests sent-requests ext-received ext-sent-requests) flatten (sort-by :mtime >))]
     {:given given
      :received received
      :requests requests
      :sent-requests sent-requests
+     :ext-received ext-received
+     :ext-sent-requests ext-sent-requests
      :all-endorsements all}))
 
 (defn all-user-endorsements-p [ctx user-id]
@@ -161,27 +175,29 @@
  [ctx user-badge-id user-id]
  (select-user-badge-endorsement-request-by-issuer-id {:user_badge_id user-badge-id :issuer_id user-id } (into {:result-set-fn first} (get-db ctx))))
 
-(defn request-endorsement! [ctx user-badge-id owner-id user-ids content]
+(defn request-endorsement! [ctx user-badge-id owner-id {:keys [user-ids emails content]}]
  (try+
   (if-not (badge-owner? ctx user-badge-id owner-id)
    (throw+ {:status "error" :message "User cannot request endorsement for a badge they do not own"})
-   (doseq [id user-ids]
-    (if-let [check (-> (request-sent? ctx user-badge-id id) :id)]
-     (throw+ {:status "error" :message "Request already sent to user"})
-     (let [endorser-info (as-> (first (plugin-fun (get-plugins ctx) "db" "user-information")) $
-                              (if $ ($ ctx id) {}))
-           {:keys [first_name last_name]} endorser-info
-           user-connection (as-> (first (plugin-fun (get-plugins ctx) "db" "get-connections-user")) $
-                                 (if $ (some-> ($ ctx owner-id id) :status) nil))
-           request-id (-> (request-endorsement<! {:id user-badge-id
-                                                  :content content
-                                                  :issuer_name (str first_name " " last_name)
-                                                  :issuer_id id
-                                                  :issuer_url (str (get-full-path ctx) "/profile/" id)}  (get-db ctx))
-                          :generated_key)]
-       (publish ctx :request_endorsement {:verb "request_endorsement" :type "badge" :subject owner-id :object request-id})
-       (when-not user-connection (as-> (first (plugin-fun (get-plugins ctx) "db" "create-connections-user!")) $   ;;create user connection if not existing
-                                       (when $ ($ ctx owner-id id))))))))
+   (do
+    (doseq [id user-ids]
+     (if-let [check (-> (request-sent? ctx user-badge-id id) :id)]
+      (throw+ {:status "error" :message "Request already sent to user"})
+      (let [endorser-info (as-> (first (plugin-fun (get-plugins ctx) "db" "user-information")) $
+                               (if $ ($ ctx id) {}))
+            {:keys [first_name last_name]} endorser-info
+            user-connection (as-> (first (plugin-fun (get-plugins ctx) "db" "get-connections-user")) $
+                                  (if $ (some-> ($ ctx owner-id id) :status) nil))
+            request-id (-> (request-endorsement<! {:id user-badge-id
+                                                   :content content
+                                                   :issuer_name (str first_name " " last_name)
+                                                   :issuer_id id
+                                                   :issuer_url (str (get-full-path ctx) "/profile/" id)}  (get-db ctx))
+                           :generated_key)]
+        (publish ctx :request_endorsement {:verb "request_endorsement" :type "badge" :subject owner-id :object request-id})
+        (when-not user-connection (as-> (first (plugin-fun (get-plugins ctx) "db" "create-connections-user!")) $   ;;create user connection if not existing
+                                        (when $ ($ ctx owner-id id)))))))
+    (when (seq emails) (ext/request-external-endorsements ctx user-badge-id owner-id emails content))))
   {:status "success"}
   (catch Object ex
     (log/error ex)
@@ -217,10 +233,12 @@
   :request (select-user-endorsement-request-status {:id user-badge-id :issuer_id user-id} (into {:result-set-fn first :row-fn :status} (get-db ctx)))})
 
 (defn pending-endorsement-count [ctx user-badge-id user-id]
- (pending-user-badge-endorsement-count {:id user-badge-id} (into {:result-set-fn first :row-fn :count} (get-db ctx))))
+ (+ (pending-user-badge-endorsement-count {:id user-badge-id} (into {:result-set-fn first :row-fn :count} (get-db ctx)))
+    (ext/pending-ext-endorsement-count {:id user-badge-id} (into {:result-set-fn first :row-fn :count} (get-db ctx)))))
 
 (defn accepted-endorsement-count [ctx user-badge-id user-id]
- {:user_endorsement_count (->> (select-accepted-badge-endorsements {:id user-badge-id}  (get-db ctx)) count)})
+ {:user_endorsement_count (+ (->> (select-accepted-badge-endorsements {:id user-badge-id}  (get-db ctx)) count)
+                             (ext/accepted-ext-endorsement-count {:id user-badge-id} (into {:result-set-fn first :row-fn :count} (get-db ctx))))})
 
 (defn endorsements-count [ctx user-badge-id user-id]
  {:pending_endorsements_count (pending-endorsement-count ctx user-badge-id user-id)