+++ src/clj/salava/oauth/db.clj
@@ -141,14 +141,15 @@
   (some->> code_verifier (u/digest "sha256") u/bytes->base64-url))
 
 ;; Expiry time for OAuth access token in seconds. Low value used for debugging.
-(def token-expires 120)
+(def token-expires 600)
 
 (defn new-access-token [ctx client_id auth_code code_verifier]
   (when-let [user (select-oauth2-auth-code-user {:client_id client_id :auth_code auth_code :auth_code_challenge (challenge-hash code_verifier)} (u/get-db-1 ctx))]
     (let [rtoken (u/random-token auth_code)
           expires token-expires]
       (update-oauth2-auth-code! {:client_id client_id :auth_code auth_code :rtoken (hashers/derive rtoken {:alg :bcrypt+sha512})} (get-db ctx))
-      {:access_token (user-session ctx user expires)
+      {:token_id (:token_id user)
+       :access_token (user-session ctx user expires)
        :refresh_token (str (:id user) "-" rtoken)
        :token_type "bearer",
        :expires expires})))
@@ -164,6 +165,10 @@
          :refresh_token (str (:id user) "-" out-token)
          :token_type "bearer"
          :expires expires}))))
+
+(defn set-firebase-token [ctx client-id user-id oauth2-id token]
+  (update-oauth2-firebase-token! {:id oauth2-id :client_id client-id :user_id user-id :token token} (get-db ctx))
+  {:success true})
 
 (defn unauthorize-client [ctx client_id user_id]
   (delete-oauth2-token! {:client_id client_id :user_id user_id} (get-db ctx))