+++ src/cljs/salava/badge/ui/modal.cljs
@@ -24,7 +24,13 @@
    [salava.core.ui.tag :as tag]
    [salava.social.ui.follow :refer [follow-badge]]
    [salava.social.ui.badge-message-modal :refer [badge-message-link]]
-   [salava.translator.ui.helper :refer [translate]]))
+   [salava.translator.ui.helper :refer [translate]]
+   [salava.badge.ui.ext-endorsement :as ext]))
+
+(defn- init-owner-profile-visibility [user-id state]
+  (ajax/GET
+   (path-for (str "/obpv1/profile/user/visibility/" user-id))
+   {:handler (fn [data] (reset! (cursor state [:profile_visibility]) data))}))
 
 (defn- init-badge-connection [state badge-id]
   (ajax/GET
@@ -45,6 +51,7 @@
 
 (defn- init-more [state]
   (init-endorsement-count (:id @state) state)
+  (init-owner-profile-visibility (:user_id @state) state)
   (when (:owner? @state)
     (init-new-message-count state)
     (ajax/GET
@@ -67,6 +74,7 @@
                                          :evidence {:url nil}
                                          :request-comment " "
                                          :selected-users []
+                                         :external-users []
                                          :request-mode false))
                     (init-more state)
                     (if (or (:user-logged-in @state) (:owner? @state))  (init-badge-connection state (:badge_id data)))))}
@@ -81,6 +89,7 @@
                             :permission "success"
                             :request-comment " "
                             :selected-users []
+                            :external-users []
                             :request-mode false))
        (init-more state)
        (if (or (:user-logged-in @state) (:owner? @state)) (init-badge-connection state (:badge_id data))))))
@@ -318,7 +327,9 @@
         (if (pos? @show-recipient-name-atom)
           (if (and user-logged-in? (not owner?))
             [:div [:span._label (t :badge/Recipient) ": "]  [:a {:href "#" :on-click #(do (.preventDefault %) (mo/open-modal [:profile :view] {:user-id owner}))} #_{:href (path-for (str "/profile/" owner))} first_name " " last_name]]
-            [:div [:span._label (t :badge/Recipient) ": "]  first_name " " last_name]))
+            (if (and (= "public" @(cursor state [:profile_visibility])) (= 2 @show-recipient-name-atom))
+              [:div [:span._label (t  :badge/Recipient) ": "] [:a.link {:href "#" :on-click #(do (.preventDefault %) (mo/open-modal [:profile :view] {:user-id owner})) #_(path-for (str "/profile/" owner))} first_name " " last_name]]
+              [:div [:span._label (t  :badge/Recipient) ": "]  first_name " " last_name])))
 
         ;metabadges
         (when owner?
@@ -442,5 +453,6 @@
            :endorse endr/endorse-badge-content
            :userbadgeendorsement endr/user-badge-endorsement-content
            :userendorsement endr/user-endorsement-content
+           :extuserendorsement ext/ext-endorsement-content
            :requestendorsement endr/request-endorsement
            :my block/mybadgesmodal}})